fix(Dockerfile): relocate generate_config.sh to /etc/prometheus for better organization
Move the `generate_config.sh` script from `/bin` to `/etc/prometheus` to better align with the convention of storing configuration-related scripts in the `/etc` directory. This change improves the organization and maintainability of the Docker image by keeping configuration scripts in a more appropriate location. Adjust the `ENTRYPOINT` and `RUN chmod` commands accordingly to reflect the new path.