chore(Dockerfile): remove chmod from COPY command and set script as entrypoint fix(generate_config.sh): change file mode to executable
Remove the `--chmod=+x` option from the COPY command in the Dockerfile
as the script's executable permission is now set directly in the
repository. Set the `generate_config.sh` script as the container
entrypoint to ensure it runs when the container starts. Add a CMD
instruction to provide default Prometheus command-line options. This
change simplifies the Dockerfile and ensures the script is executed
correctly with the necessary permissions.